pyproject.toml
[tool.poetry]
name = "chat-backend"
version = "0.1.0"
description = ""
authors = ["Bekzod Mirahmedov <notarious2@gmail.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "3.11.6"
fastapi = "^0.103.1"
pydantic = {extras = ["email"], version = "^2.3.0"}
uvicorn = "^0.23.2"
websockets = "^11.0.3"
alembic = "^1.12.0"
sqlalchemy = "^2.0.20"
pydantic-settings = "^2.0.3"
redis = "<5.0.0"
asyncpg = "^0.28.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.25.0"
pyjwt = "^2.8.0"
python-multipart = "^0.0.6"
fastapi-pagination = "^0.12.9"
fastapi-limiter = "^0.1.5"
asgi-lifespan = "^2.1.0"
sqladmin = "^0.16.0"
itsdangerous = "^2.1.2"
gunicorn = "^21.2.0"
aiofiles = "^23.2.1"
boto3 = "^1.33.9"
Pillow = "10.1.0"
sentry-sdk = {extras = ["asyncpg", "fastapi"], version = "^1.39.1"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
isort = "^5.12.0"
flake8 = "^6.1.0"
black = "^23.9.1"
flake8-pyproject = "^1.2.3"
pytest-asyncio = "^0.21.1"
pytest-env = "^1.0.1"
pytest-mock = "^3.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
skip_glob = ["**/migrations/**"]

[tool.flake8]
max-line-length = 120

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
asyncio_mode="auto"
markers = [
    "integration: Custom mark for integration tests"
]
env = [
    "ENVIRONMENT = test",
]